First Commit
This is a basic implementation of an application that plots and projects satellites and their paths based on TLE.

app.js is a stylized representation of satellite positions, limited to 5 because of rendering time.

integrated.py is a flask application that retrieves the TLE data and hosts it for the app.js to grab. IT also estimates the future flight path of the satellite.

The next big hurdle is to attach it to a backend.