Add A Postgress DB
LoadSatDb.py is still untested however, it should load a postgress db with the Caletrek API call. It has some fields that will not be loaded but those are reserved for future use.

In addition, Posgress SQL Create Tables.txt is the SQL I used to create the DB tables.